requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
climage==0.2.2
idna==3.10
kdtree==0.16
pillow==10.4.0
requests==2.32.3
termcolor==2.4.0
urllib3==2.2.3
